fix. 앱이용알림 컴포넌트화 #312
