+++ src/pages/LogInPage.tsx
@@ -1,30 +1,19 @@
-import { InstallModal } from 'components';
+import { AddToHomeScreen, InstallModal } from 'components';
 import TestLogin from 'components/TestLogin';
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { useSearchParams } from 'react-router-dom';
 import { KAKAO_OAUTH_URL } from 'shared/constants/constants';
 import { LOGO } from 'shared/constants/icons';
-import useA2hs from 'shared/hooks/useA2hs';
 import useModal from 'shared/hooks/useModal';
-import isMobile, { isInstalled, isIos } from 'shared/utils/deviceDetector';
-import { Background, Button, KakaoButton } from 'ui';
+import isMobile from 'shared/utils/deviceDetector';
+import { Background, KakaoButton } from 'ui';
 import * as t from './loginPage.style';
 
 export default function LogInPage() {
   const onLogin = () => (window.location.href = KAKAO_OAUTH_URL);
   const [searchParams] = useSearchParams();
   const channel = searchParams.get('channel');
-
-  const { deferredPrompt, installApp, clearPrompt } = useA2hs();
-  const isIOS = isIos();
   const installModal = useModal();
-  const isIosInstalled = isInstalled();
-  const [showKakaoButton, setShowKakaoButton] = useState(false);
-
-  const handleClearPromptOnAndroid = () => {
-    clearPrompt();
-    setShowKakaoButton(true);
-  };
 
   useEffect(() => {
     localStorage.setItem('invite-code', channel);
@@ -49,37 +38,7 @@ export default function LogInPage() {
         </p>
         <TestLogin />
         {isMobile ? (
-          <>
-            {isIOS && !showKakaoButton && !isIosInstalled ? (
-              <div>
-                <Button
-                  onClick={installModal.open}
-                  text="편하게 앱 이용하기"
-                  category="confirm"
-                  bgColor="#fff"
-                  color="#4059DE"
-                />
-                <button onClick={() => setShowKakaoButton(true)}>
-                  모바일 웹 이용하기
-                </button>
-              </div>
-            ) : deferredPrompt && !showKakaoButton ? (
-              <div>
-                <Button
-                  onClick={installApp}
-                  text="편하게 앱 이용하기"
-                  category="confirm"
-                  bgColor="#fff"
-                  color="#4059DE"
-                />
-                <button onClick={handleClearPromptOnAndroid}>
-                  모바일 웹 이용하기
-                </button>
-              </div>
-            ) : (
-              <KakaoButton onClick={onLogin} />
-            )}
-          </>
+          <AddToHomeScreen onLogin={onLogin} onOpen={installModal.open} />
         ) : (
           <KakaoButton onClick={onLogin} />
         )}
